Improve package structure
