+++ cli/src/main/kotlin/io/spine/protodata/cli/app/Main.kt
@@ -24,7 +24,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.protodata.cli
+package io.spine.protodata.cli.app
 
 import com.github.ajalt.clikt.completion.CompletionCandidates
 import com.github.ajalt.clikt.core.CliktCommand
@@ -46,17 +46,17 @@ import io.spine.protobuf.outerClass
 import io.spine.protodata.context.Pipeline
 import io.spine.protodata.config.Configuration
 import io.spine.protodata.config.ConfigurationFormat
-import io.spine.protodata.cli.param.ConfigFileParam
-import io.spine.protodata.cli.param.ConfigFormatParam
-import io.spine.protodata.cli.param.ConfigValueParam
-import io.spine.protodata.cli.param.OptionProviderParam
-import io.spine.protodata.cli.param.Parameter
-import io.spine.protodata.cli.param.PluginParam
-import io.spine.protodata.cli.param.RendererParam
-import io.spine.protodata.cli.param.RequestParam
-import io.spine.protodata.cli.param.SourceRootParam
-import io.spine.protodata.cli.param.TargetRootParam
-import io.spine.protodata.cli.param.UserClasspathParam
+import io.spine.protodata.cli.ConfigFileParam
+import io.spine.protodata.cli.ConfigFormatParam
+import io.spine.protodata.cli.ConfigValueParam
+import io.spine.protodata.cli.OptionProviderParam
+import io.spine.protodata.cli.Parameter
+import io.spine.protodata.cli.PluginParam
+import io.spine.protodata.cli.RendererParam
+import io.spine.protodata.cli.RequestParam
+import io.spine.protodata.cli.SourceRootParam
+import io.spine.protodata.cli.TargetRootParam
+import io.spine.protodata.cli.UserClasspathParam
 import io.spine.protodata.renderer.SourceFileSet
 import io.spine.tools.code.manifest.Version
 import java.io.File
