+++ cli/src/main/kotlin/io/spine/protodata/cli/app/ReflectiveBuilder.kt
@@ -24,7 +24,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.protodata.cli
+package io.spine.protodata.cli.app
 
 import kotlin.reflect.KClass
 import kotlin.reflect.KVisibility
