+++ cli/src/test/kotlin/io/spine/protodata/cli/app/CliTest.kt
@@ -24,7 +24,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.protodata.param
+package io.spine.protodata.cli.app
 
 import com.github.ajalt.clikt.core.MissingOption
 import com.github.ajalt.clikt.core.UsageError
@@ -35,7 +35,7 @@ import io.spine.base.Time
 import io.spine.json.Json
 import io.spine.option.OptionsProto
 import io.spine.protobuf.AnyPacker
-import io.spine.protodata.cli.Run
+import io.spine.protodata.cli.app.Run
 import io.spine.protodata.cli.given.CustomOptionPlugin
 import io.spine.protodata.cli.given.CustomOptionRenderer
 import io.spine.protodata.cli.given.DefaultOptionsCounterPlugin
